refactor: migrate entity IDs to UUIDs and remove unnecessary service implementations
