+++ src/main/java/com/WorkoutHub/workout_hub/dto/WorkoutCreationDto.java
@@ -1,9 +1,5 @@
 package com.WorkoutHub.workout_hub.dto;
 
-import com.WorkoutHub.workout_hub.entity.Exercise;
-import com.WorkoutHub.workout_hub.entity.GymRat;
-import com.WorkoutHub.workout_hub.entity.Workout;
-import com.WorkoutHub.workout_hub.entity.WorkoutPost;
 import com.WorkoutHub.workout_hub.enums.Visibility;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
@@ -12,8 +8,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.time.Duration;
-import java.time.LocalDateTime;
+import java.time.OffsetDateTime;
 import java.util.List;
 
 @Data
@@ -33,7 +28,7 @@ public class WorkoutCreationDto {
     private String caption;
 
     @NotNull
-    private LocalDateTime startTime;
+    private OffsetDateTime startTime;
 
     @NotNull
     private int duration;
