+++ src/main/java/com/WorkoutHub/workout_hub/dto/WorkoutDto.java
@@ -1,7 +1,5 @@
 package com.WorkoutHub.workout_hub.dto;
 
-import com.WorkoutHub.workout_hub.entity.Exercise;
-import com.WorkoutHub.workout_hub.entity.Workout;
 import com.WorkoutHub.workout_hub.enums.Visibility;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.AllArgsConstructor;
@@ -9,9 +7,9 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.time.Duration;
-import java.time.LocalDateTime;
+import java.time.OffsetDateTime;
 import java.util.List;
+import java.util.UUID;
 
 @Data
 @Builder
@@ -19,11 +17,11 @@ import java.util.List;
 @AllArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class WorkoutDto {
-    private int id;
+    private UUID id;
     private String workoutTitle;
-    private LocalDateTime createdAt;
+    private OffsetDateTime createdAt;
     int duration;
-    private float totalVolume;
+    private double totalVolume;
     private int numberOfSets;
     private Visibility visibility;
     private List<ExerciseDto> exercises;
