+++ src/main/java/com/WorkoutHub/workout_hub/entity/Exercise.java
@@ -4,7 +4,7 @@ import jakarta.persistence.*;
 import lombok.*;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.UUID;
+
 
 @Entity
 @Table(name = "workout_exercises")
