+++ src/main/java/com/WorkoutHub/workout_hub/entity/Routine.java
@@ -6,7 +6,6 @@ import lombok.*;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.UUID;
 
 @Entity
 @Table(name = "routines")
