+++ src/main/java/com/WorkoutHub/workout_hub/entity/RoutineExercise.java
@@ -4,7 +4,6 @@ import jakarta.persistence.*;
 import lombok.*;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.UUID;
 
 @Entity
 @Table(name = "routine_exercises")
