+++ src/main/java/com/WorkoutHub/workout_hub/entity/Workout.java
@@ -7,7 +7,6 @@ import lombok.*;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.UUID;
 
 
 @Entity
