+++ src/main/java/com/WorkoutHub/workout_hub/repository/GymRatRepo.java
@@ -6,15 +6,16 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
+import java.util.UUID;
 
 @Repository
-public interface GymRatRepo extends JpaRepository<GymRat, Integer> {
+public interface GymRatRepo extends JpaRepository<GymRat, UUID> {
     boolean existsByUsername(String username);
 
     boolean existsByEmail(String email);
 
     @Query("SELECT gr FROM GymRat gr JOIN FETCH gr.profile WHERE gr.id = :id")
-    Optional<GymRat> findGymRatWithProfileById(int id);
+    Optional<GymRat> findGymRatWithProfileById(UUID id);
 
 
     Optional<GymRat> findGymRatByUsername(String username);
