+++ src/main/java/com/WorkoutHub/workout_hub/repository/WorkoutRepo.java
@@ -7,11 +7,12 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.Query;
 
 import java.util.Optional;
+import java.util.UUID;
 
-public interface WorkoutRepo extends JpaRepository<Workout, Integer> {
+public interface WorkoutRepo extends JpaRepository<Workout, UUID> {
 
     @Query("SELECT w FROM Workout w JOIN FETCH w.exercises WHERE w.id = :id")
     Optional<Workout> findById(int id);
 
-    Page<Workout> findByGymRatId(int userId, Pageable pageable);
+    Page<Workout> findByGymRatId(UUID userId, Pageable pageable);
 }
