+++ src/main/java/com/WorkoutHub/workout_hub/service/GymRatService.java
@@ -4,15 +4,70 @@ import com.WorkoutHub.workout_hub.dto.auth.RegisterRequest;
 import com.WorkoutHub.workout_hub.dto.GymRatUpdateDto;
 import com.WorkoutHub.workout_hub.dto.ProfileDto;
 import com.WorkoutHub.workout_hub.dto.UserDto;
+import com.WorkoutHub.workout_hub.entity.GymRat;
+import com.WorkoutHub.workout_hub.entity.GymRatProfile;
+import com.WorkoutHub.workout_hub.repository.GymRatRepo;
 import jakarta.persistence.EntityNotFoundException;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.UUID;
+import java.util.stream.Collectors;
 
-public interface GymRatService {
-    List<UserDto> getAllGymRats();
-    void createGymRat(RegisterRequest creationDto);
-    UserDto getGymRatById(int id);
-    ProfileDto getGymRatProfileById(int id);
-    ProfileDto updateGymRatById(int id, GymRatUpdateDto updatedGymRat) throws EntityNotFoundException;
-    void deleteGymRatById(int id);
+@Service
+public class GymRatService {
+
+    private final GymRatRepo gymRatRepository;
+
+    @Autowired
+    public GymRatService(GymRatRepo grRepo) {
+        this.gymRatRepository = grRepo;
+    }
+
+
+    public List<UserDto> getAllGymRats() {
+        return gymRatRepository.findAll().stream()
+                .map(UserDto::new)
+                .collect(Collectors.toList());
+    }
+
+
+    public void createGymRat(RegisterRequest creationDto) {
+        GymRat gymrat = new GymRat(creationDto);
+        GymRatProfile gymRatProfile = new GymRatProfile(creationDto);
+        gymrat.setProfile(gymRatProfile);
+        gymRatRepository.save(gymrat);
+    }
+
+
+    public UserDto getGymRatById(UUID id) {
+        GymRat gymRat = gymRatRepository.findById(id)
+                 .orElseThrow(() -> new EntityNotFoundException("Entity Not Found: Gymrat with id: " + id + " is not found."));
+        return new UserDto(gymRat);
+    }
+
+
+    public ProfileDto getGymRatProfileById(UUID id) {
+        GymRat gymRat = gymRatRepository.findGymRatWithProfileById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Entity Not Found: Gymrat with id: " + id + " is not found."));
+        return new ProfileDto(gymRat);
+    }
+
+
+    public ProfileDto updateGymRatById(UUID id, GymRatUpdateDto updateDto) throws EntityNotFoundException {
+        GymRat gymrat = gymRatRepository.findGymRatWithProfileById(id)
+                .orElseThrow(() ->  new EntityNotFoundException("Entity Not Found: Gymrat with id: " + id + " is not found."));
+        gymrat.update(updateDto);
+        gymRatRepository.save(gymrat);
+        return new ProfileDto(gymrat);
+    }
+
+
+    public void deleteGymRatById(UUID id) {
+        boolean gymratExists = gymRatRepository.existsById(id);
+        if(!gymratExists)
+            throw new EntityNotFoundException("Entity Not Found: Gymrat with id: " + id + " is not found.");
+        gymRatRepository.deleteById(id);
+    }
 }
